+++ native/src/main/java/io/ballerina/stdlib/constraint/validators/StringConstraintValidator.java
@@ -20,13 +20,16 @@ package io.ballerina.stdlib.constraint.validators;
 
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
+import io.ballerina.stdlib.constraint.validators.interfaces.LengthValidator;
+import io.ballerina.stdlib.constraint.validators.interfaces.PatternValidator;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * Extern functions for validating string constraints `@constraint:String` of Ballerina.
  */
-public class StringConstraintValidator extends AbstractLengthValidator {
+public class StringConstraintValidator implements LengthValidator, PatternValidator {
 
     private final List<String> failedConstraints;
 
@@ -35,21 +38,30 @@ public class StringConstraintValidator extends AbstractLengthValidator {
     }
 
     public void validate(BMap<BString, Object> constraints, String fieldValue, String path) {
-        super.validate(constraints, fieldValue, failedConstraints, path);
+        for (Map.Entry<BString, Object> constraint : constraints.entrySet()) {
+            validate(constraint, fieldValue, failedConstraints, path);
+        }
     }
 
     @Override
-    boolean validateLength(Object fieldValue, long constraintValue) {
+    public void validate(Map.Entry<BString, Object> constraint, Object fieldValue, List<String> failedConstraints,
+                         String path) {
+        LengthValidator.super.validate(constraint, fieldValue, failedConstraints, path);
+        PatternValidator.super.validate(constraint, fieldValue, failedConstraints, path);
+    }
+
+    @Override
+    public boolean validateLength(Object fieldValue, long constraintValue) {
         return ((String) fieldValue).length() == constraintValue;
     }
 
     @Override
-    boolean validateMinLength(Object fieldValue, long constraintValue) {
+    public boolean validateMinLength(Object fieldValue, long constraintValue) {
         return ((String) fieldValue).length() >= constraintValue;
     }
 
     @Override
-    boolean validateMaxLength(Object fieldValue, long constraintValue) {
+    public boolean validateMaxLength(Object fieldValue, long constraintValue) {
         return ((String) fieldValue).length() <= constraintValue;
     }
 }