+++ native/src/main/java/io/ballerina/stdlib/constraint/validators/interfaces/ValueValidator.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.constraint.validators;
+package io.ballerina.stdlib.constraint.validators.interfaces;
 
 import io.ballerina.runtime.api.values.BString;
 
@@ -30,30 +30,30 @@ import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MIN_VALUE_EXCL
 import static io.ballerina.stdlib.constraint.Constants.SYMBOL_SEPARATOR;
 
 /**
- * The abstract class to validate the value related constraints.
+ * The interface to validate the value related constraints.
  */
-public abstract class AbstractValueValidator {
+public interface ValueValidator {
 
-    public void validate(Map.Entry<BString, Object> constraint, Number fieldValue, Number constraintValue,
-                         List<String> failedConstraints, String path) {
+    default void validate(Map.Entry<BString, Object> constraint, Object fieldValue, List<String> failedConstraints,
+                          String path) {
         switch (constraint.getKey().getValue()) {
             case CONSTRAINT_MIN_VALUE:
-                if (!validateMinValue(fieldValue, constraintValue)) {
+                if (!validateMinValue(fieldValue, constraint.getValue())) {
                     failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MIN_VALUE);
                 }
                 break;
             case CONSTRAINT_MAX_VALUE:
-                if (!validateMaxValue(fieldValue, constraintValue)) {
+                if (!validateMaxValue(fieldValue, constraint.getValue())) {
                     failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MAX_VALUE);
                 }
                 break;
             case CONSTRAINT_MIN_VALUE_EXCLUSIVE:
-                if (!validateMinValueExclusive(fieldValue, constraintValue)) {
+                if (!validateMinValueExclusive(fieldValue, constraint.getValue())) {
                     failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MIN_VALUE_EXCLUSIVE);
                 }
                 break;
             case CONSTRAINT_MAX_VALUE_EXCLUSIVE:
-                if (!validateMaxValueExclusive(fieldValue, constraintValue)) {
+                if (!validateMaxValueExclusive(fieldValue, constraint.getValue())) {
                     failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MAX_VALUE_EXCLUSIVE);
                 }
                 break;
@@ -62,11 +62,11 @@ public abstract class AbstractValueValidator {
         }
     }
 
-    abstract boolean validateMinValue(Number fieldValue, Number constraintValue);
+    boolean validateMinValue(Object fieldValue, Object constraintValue);
 
-    abstract boolean validateMaxValue(Number fieldValue, Number constraintValue);
+    boolean validateMaxValue(Object fieldValue, Object constraintValue);
 
-    abstract boolean validateMinValueExclusive(Number fieldValue, Number constraintValue);
+    boolean validateMinValueExclusive(Object fieldValue, Object constraintValue);
 
-    abstract boolean validateMaxValueExclusive(Number fieldValue, Number constraintValue);
+    boolean validateMaxValueExclusive(Object fieldValue, Object constraintValue);
 }